Update 'Admin.vue' and 'AddFeed.vue' to handle prop options
The 'Admin.vue' and 'AddFeed.vue' have been updated to define and handle props. Specifically, new properties `options` and `clubs` are defined in 'Admin.vue' and an `options` property is parsed and used in 'AddFeed.vue'. This update enhances functionality by allowing additional data to be transferred and used within these components.