Add title attribute to feed and utilize it in URLs
Added a new 'title' attribute to the feeds and provided support for it to be created and edited in the admin section. Replaced using the 'id' with the 'slug' in URL paths for feeds for better readability and SEO capability. This is reflected in changes made to multiple files including EditFeed.vue, NewFeed.vue, ListFeed.vue, and others. Also created a new Vue component for admin navigation and performed minor json format fixes.