Feature/admin add option 01 (#753)
* Add new 'AddFeed' component to Admin.vue

The commit introduces a new 'AddFeed' component, which is added to 'Admin.vue'. This new component contains a 'slug' input field in its template, allowing administrators to add new feeds. It expands the admin section functionality to include feed management options.

* Update 'Admin.vue' and 'AddFeed.vue' to handle prop options

The 'Admin.vue' and 'AddFeed.vue' have been updated to define and handle props. Specifically, new properties `options` and `clubs` are defined in 'Admin.vue' and an `options` property is parsed and used in 'AddFeed.vue'. This update enhances functionality by allowing additional data to be transferred and used within these components.

* Refactor feed management components in Admin section

This commit introduces changes in the Admin section related to feed management. It removes the 'AddFeed' component and replaces it with three new components: 'EditFeed', 'ListFeed', and 'NewFeed'. In doing so, it enhances the feed management flexibility and allows for feed editing and listing, in addition to creating new feeds. The components are wired into the respective new pages in the application. The overall aim is to provide administrators with more control over feed management.

* Refactor code for feed handling in Admin module

This commit streamlines code by moving the 'feeds' definitions from the 'Admin.vue' to respective 'index.ts' and 'ListFeed.vue'. Unnecessary comments and methods related to 'feeds' in 'Admin.vue' and 'ListFeed.vue' were also removed. Moreover, an 'EditFeed' component is introduced to enable feed editing functionality and routing has been adjusted accordingly in 'index.ts' and 'ListFeed.vue'. More precise props definition is implemented in 'EditFeed.vue' providing a cleaner way to handle data.

* Add title attribute to feed and utilize it in URLs

Added a new 'title' attribute to the feeds and provided support for it to be created and edited in the admin section. Replaced using the 'id' with the 'slug' in URL paths for feeds for better readability and SEO capability. This is reflected in changes made to multiple files including EditFeed.vue, NewFeed.vue, ListFeed.vue, and others. Also created a new Vue component for admin navigation and performed minor json format fixes.

* Implement error handling for feed slug input

Added checks for empty or duplicate slug values in the feed creation and editing process. The changes involve indicating an error state and displaying appropriate error messages in the NewFeed.vue and EditFeed.vue components. This improves the user's understanding of input requirements and enhances the overall application robustness.

* tweaking

* restructuring

---------

Co-authored-by: Kazuyoshi Kawakami <saruyama.monki@gmail.com>
Co-authored-by: aggre <hiroyuki.aggre@gmail.com>